+++ src/vue/forms/GeneralKycForm.vue
@@ -78,15 +78,13 @@ import { api } from '@/api'
 import { ErrorHandler } from '@/js/helpers/error-handler'
 
 import { required } from '@validators'
-import { mapGetters, mapActions } from 'vuex'
+import { mapGetters } from 'vuex'
 import { vuexTypes } from '@/vuex'
 import { vueRoutes } from '@/vue-router/routes'
-import { REQUEST_STATES_STR } from '@/js/const/request-states.const'
 
 const EVENTS = {
   submitted: 'submitted',
   logout: 'logout',
-  kycRecoverySubmit: 'kyc-recovery-submit',
 }
 
 export default {
@@ -115,9 +113,6 @@ export default {
       isAccountRoleReseted: vuexTypes.isAccountRoleReseted,
       accountRoleToSet: vuexTypes.kycAccountRoleToSet,
       previousAccountRole: vuexTypes.kycPreviousRequestAccountRoleToSet,
-      kycRecoveryBlobId: vuexTypes.kycRecoveryBlobId,
-      kycRecoveryState: vuexTypes.kycRecoveryState,
-      kycRecoveryRequestBlob: vuexTypes.kycRecoveryRequestBlob,
     }),
 
     isFormPopulatable () {
@@ -132,21 +127,12 @@ export default {
   },
 
   async created () {
-    if (
-      this.isKycRecoveryPage &&
-      this.kycRecoveryBlobId &&
-      (this.kycRecoveryState !== REQUEST_STATES_STR.permanentlyRejected)
-    ) {
-      this.form = this.parseKycData(this.kycRecoveryRequestBlob)
-    } else if (!this.isKycRecoveryPage && this.isFormPopulatable) {
+    if (this.isFormPopulatable) {
       this.form = this.parseKycData(this.kycLatestRequestData)
     }
   },
 
   methods: {
-    ...mapActions({
-      sendKycRecoveryRequest: vuexTypes.SEND_KYC_RECOVERY_REQUEST,
-    }),
     tryToSubmit () {
       if (!this.isFormValid()) return
       if (this.isSignUpKycPage) {
@@ -160,17 +146,13 @@ export default {
       this.isFormSubmitting = true
       try {
         const kycBlobId = await this.createKycBlob(BLOB_TYPES.kycGeneral)
-        if (this.isKycRecoveryPage) {
-          await this.sendKycRecoveryRequest(kycBlobId)
-          this.$emit(EVENTS.kycRecoverySubmit)
-        } else {
-          const operation = this.createKycOperation(
-            kycBlobId,
-            this.kvEntryGeneralRoleId,
-          )
-          await api.postOperations(operation)
-          await this.loadKyc()
-        }
+
+        const operation = this.createKycOperation(
+          kycBlobId,
+          this.kvEntryGeneralRoleId,
+        )
+        await api.postOperations(operation)
+        await this.loadKyc()
         this.$emit(EVENTS.submitted)
       } catch (e) {
         ErrorHandler.process(e)